+++ shardis-ui/config/webpack.prod.config.js
@@ -30,6 +30,7 @@ module.exports = validateConfig(webpackMerge(baseWebpackConfig, {
     cache: false
   },
   entry: {
+    'styles': './src/main/frontend/scss/main.scss',
     'polyfills': './src/main/frontend/polyfills.ts',
     'vendor': './src/main/frontend/vendor.ts',
     'main': './src/main/frontend/main.ts'